+++ HW 4.3/DATA.txt
@@ -8,3 +8,5 @@ Sofia
 79051235566
 Diana
 71112223344
+Vadik
+920385928
