+++ HW 4.3/HW 4.3.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30611.23
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "HW 4.3", "HW 4.3.vcxproj", "{BD99415C-655A-436A-966E-71619A41F241}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BD99415C-655A-436A-966E-71619A41F241}.Debug|x64.ActiveCfg = Debug|x64
+		{BD99415C-655A-436A-966E-71619A41F241}.Debug|x64.Build.0 = Debug|x64
+		{BD99415C-655A-436A-966E-71619A41F241}.Debug|x86.ActiveCfg = Debug|Win32
+		{BD99415C-655A-436A-966E-71619A41F241}.Debug|x86.Build.0 = Debug|Win32
+		{BD99415C-655A-436A-966E-71619A41F241}.Release|x64.ActiveCfg = Release|x64
+		{BD99415C-655A-436A-966E-71619A41F241}.Release|x64.Build.0 = Release|x64
+		{BD99415C-655A-436A-966E-71619A41F241}.Release|x86.ActiveCfg = Release|Win32
+		{BD99415C-655A-436A-966E-71619A41F241}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {5CC11B2B-4BDD-4E2A-B5A4-AB746CD8DD6B}
+	EndGlobalSection
+EndGlobal
